refactor: extract constants and helper methods
