+++ src/library/calc/constants.js
@@ -1,0 +1,3 @@
+export default Object.freeze({
+  TOTAL_MONTHS_PER_YEAR: 12,
+});
