+++ src/library/calc/solver.js
@@ -1,0 +1,14 @@
+import nerdamer from 'nerdamer/nerdamer.core';
+import Algebra from 'nerdamer/Algebra';
+import Calculus from 'nerdamer/Calculus';
+import Solve from 'nerdamer/Solve';
+
+export const solveForOneUnknownVariable = (
+  equationString,
+  knownValues,
+  unknownVariable,
+) => {
+  const equation = nerdamer(equationString).evaluate(knownValues);
+  const solution = equation.solveFor(unknownVariable);
+  return Number(solution.text());
+};
